mysql upstart workaround
The problem?

https://wiki.ubuntu.com/LucidLynx/ReleaseNotes

"""Upstart jobs cannot be run in a chroot:

Upstart jobs cannot be started in a chroot because upstart acts as a
service supervisor, and processes within the chroot are unable to
communicate with the upstart running outside of the chroot (430224).
This will cause some packages that have been converted to use upstart
jobs instead of init scripts to fail to upgrade within a chroot. Users
are advised to configure their chroots with /sbin/initctl pointing to
/bin/true, with the following commands run within the chroot:

dpkg-divert --local --rename --add /sbin/initctl
ln -s /bin/true /sbin/initctl"""

https://bugs.launchpad.net/ubuntu/+source/upstart/+bug/430224

In other words, starting an upstart job inside a chroot actually starts
it outside the chroot. Upstart uses dbus to communicate with processes,
and dbus uses unix sockets in the abstract namespace, which transcends
the filesystem, and thus, chroot.